fix: input group disabled + list page complete + details page route made
