+++ ssr/src/page/mod.rs
@@ -6,5 +6,9 @@ pub use root::*;
 mod hotel_details;
 pub use hotel_details::*;
 
+
+mod hotel_list;
+pub use hotel_list::*;
+
 pub mod location_search;
 // pub use location_search::*;
